database: Remember versions and formats of all temporary TOC files.
The patch fixes a bug introduce by commit 089b54f2d2356b8b35d6b164cdf8f884216e0910.
This bug exhibited when master was deployed in an attempt to populate
materialised views. The nodes restarted in the middle and they were not able
to come back.

The fix is to remember formats and versions of sstables for every generation.

Fixes: #3324.

Signed-off-by: Daniel Fiala <daniel@scylladb.com>
Message-Id: <20180410083114.17315-1-daniel@scylladb.com>